pyproject.toml
[tool.poetry]
name = "drf-doc-demo"
version = "0.1.0"
description = ""
authors = ["shihono <26952997+shihono@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
drf-spectacular = "^0.24.1"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
alphabet2kana = "^0.1.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

